// packages/astro-loader-github-releases/src/graphql/gen/operations.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
// ...
export type Scalars = {
  ID: { input: string; output: string }
  String: { input: string; output: string }
  Boolean: { input: boolean; output: boolean }
  Int: { input: number; output: number }
  Float: { input: number; output: number }
  /** A (potentially binary) string encoded using base64. */
  Base64String: { input: any; output: any }
  /**
   * Represents non-fractional signed whole numeric values. Since the value may
   * exceed the size of a 32-bit integer, it's encoded as a string.
   */
  BigInt: { input: any; output: any }
  /** An ISO-8601 encoded date string. */
  Date: { input: any; output: any }
  /** An ISO-8601 encoded UTC date string. */
  DateTime: { input: string; output: string }
  /** A Git object ID. */
  GitObjectID: { input: any; output: any }
  /** A fully qualified reference name (e.g. `refs/heads/master`). */
  GitRefname: { input: any; output: any }
  /** Git SSH string */
  GitSSHRemote: { input: any; output: any }
  /** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
  GitTimestamp: { input: any; output: any }
  /** A string containing HTML code. */
  HTML: { input: string; output: string }
  /** An ISO-8601 encoded UTC date string with millisecond precision. */
  PreciseDateTime: { input: any; output: any }
  /** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
  URI: { input: string; output: string }
  /** A valid x509 certificate string */
  X509Certificate: { input: any; output: any }
}
// ...
export type GetReleasesQueryVariables = Exact<{
  owner: Scalars['String']['input']
  repo: Scalars['String']['input']
  first: Scalars['Int']['input']
  cursor?: InputMaybe<Scalars['String']['input']>
}>

export type GetReleasesQuery = {
  repository?: {
    releases: {
      nodes?: Array<{
        id: string
        url: string
        name?: string | null
        tagName: string
        description?: string | null
        descriptionHTML?: string | null
        isDraft: boolean
        isLatest: boolean
        isPrerelease: boolean
        createdAt: string
        publishedAt?: string | null
        repository: {
          name: string
          nameWithOwner: string
          url: string
          stargazerCount: number
          isInOrganization: boolean
        }
      } | null> | null
      pageInfo: { hasNextPage: boolean; endCursor?: string | null }
    }
  } | null
}